requirements.txt
streamlit
sentence-transformers
faiss-cpu
moviepy
pydub
python-dotenv
google-generativeai
pytest
